requirements.txt
disnake==2.9.1
transformers==4.37.2
tensorflow==2.15.0
pymongo==3.12.0
python-dotenv
